custom element refactor for import export

+++ backend/app/schemacms/pages/serializers.py
@@ -359,7 +359,9 @@ class PageBaseSerializer(CustomModelSerializer):
             if element_type == constants.ElementType.CUSTOM_ELEMENT:
                 delete_elements_sets = element.pop("delete_elements_sets", [])
 
-                obj.delete_custom_elements_sets(delete_elements_sets)
+                if delete_elements_sets:
+                    obj.delete_custom_elements_sets(delete_elements_sets)
+
                 validated_element_value = self.validate_custom_element_sets(element_value)
                 obj.update_or_create_custom_element_sets(validated_element_value)
